# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AFRICA X1 NFT Staking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WalletDashboard {
  walletAddress: string;
  nftsOwned: number;
  nftsStaked: number;
  /** Pending XNT rewards (as string to preserve precision) */
  pendingXnt: string;
  pendingX1Brains: string;
  pendingAf: string;
  lifetimeXnt: string;
  lifetimeX1Brains: string;
  lifetimeAf: string;
  totalValueEarned: string;
}

export type NFTItemRarity = typeof NFTItemRarity[keyof typeof NFTItemRarity];


export const NFTItemRarity = {
  Common: 'Common',
  Uncommon: 'Uncommon',
  Rare: 'Rare',
  Epic: 'Epic',
  Legendary: 'Legendary',
} as const;

export type NFTItemStatus = typeof NFTItemStatus[keyof typeof NFTItemStatus];


export const NFTItemStatus = {
  Unstaked: 'Unstaked',
  Staked: 'Staked',
  Locked: 'Locked',
} as const;

export interface NFTItem {
  tokenId: string;
  name: string;
  imageUrl: string;
  rarity: NFTItemRarity;
  isStaked: boolean;
  status: NFTItemStatus;
  dailyRewardXnt?: string;
  dailyRewardX1Brains?: string;
  dailyRewardAf?: string;
  totalEarnedXnt?: string;
  totalEarnedX1Brains?: string;
  totalEarnedAf?: string;
  /** @nullable */
  stakedAt?: string | null;
  /** @nullable */
  pendingXnt?: string | null;
  /** @nullable */
  pendingX1Brains?: string | null;
  /** @nullable */
  pendingAf?: string | null;
}

export type StakedNFTDetailRarity = typeof StakedNFTDetailRarity[keyof typeof StakedNFTDetailRarity];


export const StakedNFTDetailRarity = {
  Common: 'Common',
  Uncommon: 'Uncommon',
  Rare: 'Rare',
  Epic: 'Epic',
  Legendary: 'Legendary',
} as const;

export interface StakedNFTDetail {
  id: number;
  tokenId: string;
  walletAddress: string;
  name: string;
  imageUrl: string;
  rarity: StakedNFTDetailRarity;
  stakedAt: string;
  pendingXnt: string;
  pendingX1Brains: string;
  pendingAf: string;
  totalEarnedXnt: string;
  totalEarnedX1Brains: string;
  totalEarnedAf: string;
  dailyRewardXnt: string;
  dailyRewardX1Brains: string;
  dailyRewardAf: string;
}

export type RarityRewardRateRarity = typeof RarityRewardRateRarity[keyof typeof RarityRewardRateRarity];


export const RarityRewardRateRarity = {
  Common: 'Common',
  Uncommon: 'Uncommon',
  Rare: 'Rare',
  Epic: 'Epic',
  Legendary: 'Legendary',
} as const;

export interface RarityRewardRate {
  rarity: RarityRewardRateRarity;
  xntPerDay: string;
  x1BrainsPerDay: string;
  afPerDay: string;
}

export interface RewardRates {
  rates: RarityRewardRate[];
  lastUpdated: string;
}

export interface StakeInput {
  walletAddress: string;
  tokenId: string;
  /** Wallet signature proving ownership */
  signature: string;
  /** Message that was signed */
  message: string;
}

export interface UnstakeInput {
  walletAddress: string;
  tokenId: string;
  signature: string;
  message: string;
}

export interface ClaimInput {
  walletAddress: string;
  tokenId: string;
  signature: string;
  message: string;
}

export interface ClaimResult {
  success: boolean;
  claimedXnt: string;
  claimedX1Brains: string;
  claimedAf: string;
  txHash: string;
}

export interface RewardSummary {
  walletAddress: string;
  totalPendingXnt: string;
  totalPendingX1Brains: string;
  totalPendingAf: string;
  stakedNftCount: number;
}

export type StakingEventEventType = typeof StakingEventEventType[keyof typeof StakingEventEventType];


export const StakingEventEventType = {
  stake: 'stake',
  unstake: 'unstake',
  claim: 'claim',
} as const;

export interface StakingEvent {
  id: number;
  eventType: StakingEventEventType;
  walletAddress: string;
  tokenId: string;
  txHash: string;
  /** @nullable */
  xntAmount?: string | null;
  /** @nullable */
  x1BrainsAmount?: string | null;
  /** @nullable */
  afAmount?: string | null;
  createdAt: string;
}

export interface StakingConfig {
  stakingEnabled: boolean;
  stakingPaused: boolean;
  claimingPaused: boolean;
  collectionAddress: string;
  supportedTokens: string[];
  updatedAt: string;
}

export interface StakingConfigUpdate {
  stakingEnabled?: boolean;
  stakingPaused?: boolean;
  claimingPaused?: boolean;
  collectionAddress?: string;
  supportedTokens?: string[];
  rewardRates?: RarityRewardRate[];
}

export interface StakingStats {
  totalStaked: number;
  totalWallets: number;
  totalXntDistributed: string;
  totalX1BrainsDistributed: string;
  totalAfDistributed: string;
  stakingEnabled: boolean;
  claimingPaused: boolean;
}

export interface TreasuryBalance {
  xntBalance: string;
  x1BrainsBalance: string;
  afBalance: string;
  lastUpdated: string;
}

export interface RewardHistoryItem {
  id: number;
  walletAddress: string;
  tokenId: string;
  xntAmount: string;
  x1BrainsAmount: string;
  afAmount: string;
  claimedAt: string;
}