# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AFRICA X1 NFT Staking API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimInput {
  walletAddress: string;
  tokenId: string;
  signature: string;
  message: string;
}