# lib/api-zod/src/generated/types/rewardHistoryItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AFRICA X1 NFT Staking API
 * OpenAPI spec version: 0.1.0
 */

export interface RewardHistoryItem {
  id: number;
  walletAddress: string;
  tokenId: string;
  xntAmount: string;
  x1BrainsAmount: string;
  afAmount: string;
  claimedAt: Date;
}